feat(defaults): add default basket config
