+++ tests/test_basket.py
@@ -1,24 +1,6 @@
 import pytest
 
-from widget_saler.basket import Basket, BasketConfig
-
-
-@pytest.fixture
-def basket_config() -> BasketConfig:
-    return BasketConfig(
-        products=[
-            {"name": "Red Widget", "code": "R01", "price": 32.95},
-            {"name": "Green Widget", "code": "G01", "price": 24.95},
-            {"name": "Blue Widget", "code": "B01", "price": 7.95},
-        ],
-        delivery_cost_rules=[
-            {"lower_than_threshold": 50, "delivery_cost": 4.95},
-            {"lower_than_threshold": 90, "delivery_cost": 2.95},
-        ],
-        special_offer_rules=[
-            {"product_code": "R01", "product_amount": 2, "discount_ratio": 0.25}
-        ],
-    )
+from widget_saler.basket import Basket
 
 
 @pytest.mark.parametrize(
@@ -31,10 +13,8 @@ def basket_config() -> BasketConfig:
         ((), 0),
     ),
 )
-def test_basket_total(
-    product_codes: list[str], expected_total: float, basket_config: BasketConfig
-):
-    basket = Basket(basket_config)
+def test_basket_total(product_codes: list[str], expected_total: float):
+    basket = Basket()
     for product_code in product_codes:
         basket.add(product_code)
